+++ MK4duo/src/commands/gcode/bedlevel/m420.h
@@ -209,7 +209,7 @@ EXIT_M420:
     if (bedlevel.z_fade_height > 0.0)
       SERIAL_EV(bedlevel.z_fade_height);
     else
-      SERIAL_EM(MSG_OFF);
+      SERIAL_EM(MSG_LCD_OFF);
   #endif
 
   // Report change in position