+++ MK4duo/src/commands/gcode/config/m305.h
@@ -122,7 +122,7 @@ inline void gcode_M305(void) {
 
   if (parser.seen('P')) {
     // Put off the heaters
-    act->setTarget(0);
+    act->set_target_temp(0);
 
     const pin_t new_pin = parser.analog_value_pin();
     if (new_pin != NoPin) {