+++ MK4duo/src/commands/gcode/temperature/m191.h
@@ -41,7 +41,7 @@ inline void gcode_M191(void) {
   if (WITHIN(c, 0 , CHAMBERS - 1)) {
     bool no_wait_for_cooling = parser.seen('S');
     if (no_wait_for_cooling || parser.seen('R'))
-      chambers[c].setTarget(parser.value_celsius());
+      chambers[c].set_target_temp(parser.value_celsius());
     else return;
 
     lcdui.set_status_P(chambers[c].isHeating() ? PSTR(MSG_CHAMBER_HEATING) : PSTR(MSG_CHAMBER_COOLING));