+++ Documentation/changelog.md
@@ -15,6 +15,7 @@
 * Add Option for Safety Timer in configuration_temperature.h
 * Add G34 and M422 Z Steppers Auto-Alignment (Cartesian and CORE)
 * Add M86 M[min] set safety timer expiration time in minute. M86 M0 will disable safety timer
+* Add M16 Expected printer check
 * Add M217 - Set Park position and tool change parameters
 *   S[linear]   Swap length
 *   E[linear]   Purge length